MODIFY: head.hpp file

+++ head.hpp
@@ -5,7 +5,7 @@
 int askBudget();
 double dailyExpLog();
 void weekLog();
-void warning(int bud,int exp);
+void warning(double bud,double exp);
 void dispHistory();
 void insights();
 void menu();
